net_localization/requirements_ml.txt
# ML training and inference dependencies
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
albumentations>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
scipy>=1.11.0
numpy>=1.24.0
